Move trailer related iterators to new module
