+++ lib/src/trailer/mod.rs
@@ -15,6 +15,7 @@
 //!
 
 pub mod accumulation;
+pub mod iter;
 pub mod spec;
 
 use regex::Regex;
@@ -155,79 +156,6 @@ impl FromStr for Trailer {
             _ => Err(Error::from_kind(EK::TrailerFormatError(s.to_owned())))
         }
     }
-}
-
-
-/// Iterator assembling trailers from key-value pairs
-///
-/// This iterator wraps an iterator returning key-value pairs. The pairs
-/// returned by the wrapped iterator are assembled to `Trailer`s.
-///
-pub struct PairsToTrailers<K, I>
-    where K: Into<TrailerKey>,
-          I: Iterator<Item = (K, TrailerValue)>
-{
-    inner: I
-}
-
-impl<K, I, J> From<J> for PairsToTrailers<K, I>
-    where K: Into<TrailerKey>,
-          I: Iterator<Item = (K, TrailerValue)>,
-          J: IntoIterator<Item = (K, TrailerValue), IntoIter = I>
-{
-    fn from(iter: J) -> Self {
-        PairsToTrailers { inner: iter.into_iter() }
-    }
-}
-
-impl<K, I> Iterator for PairsToTrailers<K, I>
-    where K: Into<TrailerKey>,
-          I: Iterator<Item = (K, TrailerValue)>
-{
-    type Item = Trailer;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        self.inner
-            .next()
-            .map(|(k, v)| Trailer { key: k.into(), value: v })
-    }
-}
-
-
-/// Iterator extracting DIT trailers from an iterator over trailers
-///
-pub struct DitTrailers<I>(I)
-    where I: Iterator<Item = Trailer>;
-
-impl<I> From<I> for DitTrailers<I>
-    where I: Iterator<Item = Trailer>
-{
-    fn from(inner: I) -> Self {
-        DitTrailers(inner)
-    }
-}
-
-impl<I> Iterator for DitTrailers<I>
-    where I: Iterator<Item = Trailer>
-{
-    type Item = Trailer;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        loop {
-            match self.0.next() {
-                None => return None,
-                Some(trailer) => {
-                    if trailer.key.0.starts_with("Dit") {
-                        return Some(trailer);
-                    } else {
-                        continue;
-                    }
-
-                }
-            }
-        }
-    }
-
 }
