+++ src/main.rs
@@ -32,7 +32,8 @@ use git2::Commit;
 use libgitdit::issue::IssueRefType;
 use libgitdit::message::LineIteratorExt;
 use libgitdit::trailer::accumulation::{self, Accumulator};
-use libgitdit::trailer::{Trailer, PairsToTrailers};
+use libgitdit::trailer::iter::PairsToTrailers;
+use libgitdit::trailer::Trailer;
 use libgitdit::{Issue, Message, RemoteExt, RepositoryExt};
 use log::LogLevel;
 use std::fs::File;
